+++ project_template/src/modules/welcome/containers/WelcomePage.js
@@ -12,7 +12,9 @@ export class WelcomePage extends Component {
     return (
       <Page>
         <Header>
-          <h1>##APP_NAME## {__VERSION_NUMBER__}</h1>
+          <h1>
+            {"<$= appName $>"} {__VERSION_NUMBER__}
+          </h1>
         </Header>
         <Main>
           <p>{I18n.t("welcome.hello")}</p>